requirements.txt
Django==1.10.5
Pillow==4.0.0
appdirs==1.4.0
django-braces==1.8.1
django-cors-headers==1.1.0
django-extensions==1.7.6
django-filter==1.0.1
django-oauth-toolkit==0.10.0
django-registration==2.0.4
django-simple-history==1.8.0
django-versatileimagefield==1.6.3
djangorestframework==3.5.4
geopy==1.11.0
oauthlib==1.0.3
olefile==0.44
packaging==16.8
psycopg2==2.6.2
py==1.4.30
pymongo==3.4.0
pyparsing==2.1.10
pytest==2.8.2
requests==2.8.1
six==1.10.0
django-filters
